api/requirements.txt
arcade==1.3.4
asn1crypto==0.24.0
boto3==1.9.28
botocore==1.12.28
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
dialogflow==0.5.1
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.6
Flask-SQLAlchemy==2.3.2
future==0.16.0
google-api-core==1.4.0
google-auth==1.5.1
googleapis-common-protos==1.5.3
googlemaps==3.0.2
grpcio==1.15.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jwt==0.5.4
MarkupSafe==1.0
numpy==1.15.0
pandas==0.23.4
Pillow==5.2.0
protobuf==3.6.1
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pyglet==1.3.2
PyJWT==1.6.4
python-dateutil==2.7.3
python-google-places==1.4.1
pytz==2018.5
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.11
torch==0.4.1
torchvision==0.2.1
urllib3==1.23
Werkzeug==0.14.1
